Message Log Base R6 Class

An abstract base class defining the interface for message logging in HydraR. Subclasses provide concrete storage implementations (Memory, File, Database).

Value

A MessageLog base object.

Methods

Public methods

MessageLog$log()

MessageLog$get_all()

MessageLog$clone()

Method log()

Store a message.

Usage

MessageLog$log(msg)

Arguments

msg

List. Message object.

Returns

The log object (invisibly).

Method get_all()

Get all logs.

Usage

MessageLog$get_all()

Returns

List of logs.

Method clone()

The objects of this class are cloneable with this method.

Usage

MessageLog$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # \dontrun{
# 1. Abstract interface usage (internal)
# Message logs are passed to dag_create() for audit trails.
dag <- dag_create(
  message_log = DuckDBMessageLog$new(db_path = "audit.duckdb")
)
} # }